Portfolio update complete without multiple image

+++ app/Http/Controllers/PortfolioController.php
@@ -40,7 +40,7 @@ class PortfolioController extends Controller
      */
     public function store(Request $request)
     {
-
+        // return request('our_role');
         $request->validate([
             'title' => 'required',
             'description' => 'required',
@@ -54,40 +54,32 @@ class PortfolioController extends Controller
             'image' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
         ]);
 
+        $portfolio = Portfolio::create([
+            'title' => $request->title,
+            'image' => 'image',
+            'title_slug' =>Str::slug($request->title),
+            'description' => $request->description,
+            'live_link' => $request->live_link,
+            'bahance_link' => $request->bahance_link,
+            'project_length' => $request->project_length,
+            'our_role' => $request->our_role,
+            'tool_used' => $request->tool_used,
+            'client_name' => $request->client_email,
+            'client_email' => $request->client_email,
+            'created_at' => Carbon::now(),
+        ]);
+
         if($request->has('image')) {
-
-
             $image = $request->image;
             $imageName = time() . '_' . uniqid() .'.'. $image->getClientOriginalExtension();
-            Storage::putFileAs('/portfolio', $image, $imageName);
 
-            $image_address = 'storage/portfolio/'. $imageName;
-
-            Portfolio::insert([
-                'title' => $request->title,
-                'image' =>$image_address,
-                'title_slug' =>Str::slug($request->title),
-                'description' => $request->description,
-                'live_link' => $request->live_link,
-                'bahance_link' => $request->bahance_link,
-                'project_length' => $request->project_length,
-                'our_role' => $request->our_role,
-                'tool_used' => $request->tool_used,
-                'client_name' => $request->client_email,
-                'client_email' => $request->client_email,
-                'created_at' => Carbon::now()
-            ]);
-
-            return back()->with('insert', 'Portfolio added Successfully');
-
-        }else {
-
-            return back()->with('error', 'Something went wrong!');
-
+            Storage::putFileAs('portfolio', $image, $imageName);
+            $portfolio->image = 'storage/portfolio/'. $imageName;
+            $portfolio->save();
         }
 
-
-
+        // Session::flash('success', '');
+        return back()->with('insert', 'Portfolio added Successfully');
     }
 
     /**
@@ -123,7 +115,6 @@ class PortfolioController extends Controller
     public function update(Request $request,$id)
     {
 
-
         $request->validate([
             'title' => 'required',
             'description' => 'required',
@@ -138,33 +129,39 @@ class PortfolioController extends Controller
 
         if($request->has('image')) {
 
-            // old image delete
-            $old_image = Portfolio::findOrFail($id);
-            unlink(base_path('public/'.$old_image->image));
+                // old image delete
+                $old_image = Portfolio::findOrFail($id);
+
+                if (file_exists($old_image->image)) {
+                    unlink(base_path('public/'.$old_image->image));
+                }
+
+                    $image = $request->image;
+                    $imageName = time() . '_' . uniqid() .'.'. $image->getClientOriginalExtension();
+                    Storage::putFileAs('/portfolio', $image, $imageName);
+                    $image_address = 'storage/portfolio/'. $imageName;
+
+                    Portfolio::findOrFail($id)->update([
+                        'title' => $request->title,
+                        'image' =>$image_address,
+                        'title_slug' =>Str::slug($request->title),
+                        'description' => $request->description,
+                        'live_link' => $request->live_link,
+                        'bahance_link' => $request->bahance_link,
+                        'project_length' => $request->project_length,
+                        'our_role' => $request->our_role,
+                        'tool_used' => $request->tool_used,
+                        'client_name' => $request->client_email,
+                        'client_email' => $request->client_email,
+                        'updated_at' => Carbon::now()
+                    ]);
 
 
-            $image = $request->image;
-            $imageName = time() . '_' . uniqid() .'.'. $image->getClientOriginalExtension();
-            Storage::putFileAs('/portfolio', $image, $imageName);
-            $image_address = 'storage/portfolio/'. $imageName;
-
-           Portfolio::findOrFail($id)->update([
-                'title' => $request->title,
-                'image' =>$image_address,
-                'title_slug' =>Str::slug($request->title),
-                'description' => $request->description,
-                'live_link' => $request->live_link,
-                'bahance_link' => $request->bahance_link,
-                'project_length' => $request->project_length,
-                'our_role' => $request->our_role,
-                'tool_used' => $request->tool_used,
-                'client_name' => $request->client_email,
-                'client_email' => $request->client_email,
-                'updated_at' => Carbon::now()
-            ]);
+                    return redirect('/portfolio')->with('update', 'Portfolio Updated Successfully With Image');
 
 
-            return redirect('/portfolio')->with('update', 'Portfolio Updated Successfully With Image');
+
+
 
         }else {
 
@@ -203,9 +200,12 @@ class PortfolioController extends Controller
 
         if($old_image){
             unlink(base_path('public/'.$old_image->image));
-            $old_image->delete();
         }
 
-       return redirect()->back()->with('delete', 'Testimonial Successfully Delete');
+        if($old_image){
+            $old_image->delete();
+            return redirect()->back()->with('delete', 'Testimonial Successfully Delete');
+        }
+
     }
 }
